// src/app/shared/config/model/stock-available.model.ts
export class StockAvailable extends BaseModel {
  // branch: Branch;
  todaysDate: Date;
  openingStock: number;
  receivedFromQuarantine: number;
  totalAvailable: number;
  totalReceipts: number;
  hospitals: number;
  receicedFromQuarantine: number;
  issueToCompats: number;
  expired: number;
  disasters: number;
  haemolysed_clots_other: number;
  wholeBloodToPackedCells: number;
  totalIssues: number;
  totalHospitalOrders: number;
  totalTransfersToOtherBranches: number;
  currentAvailableStock:  number;
  rhPositiveWbO: number;
  rhPositivePcO: number;
  rhPositivePaedWbO: number;
  rhPositivePaedPcO: number;
  rhNegativeWbO: number;
  rhNegativePcO: number;
  rhNegativePaedWbO: number;
  rhNegativePaedPcO: number;
  totalO: number;
  percentagefTotalO: number;
  rhPositiveWbA: number;
  rhPositivePcA: number;
  rhPositivePaedWbA: number;
  rhPositivePaedPcA: number;
  rhNegativeWbA: number;
  rhNegativePcA: number;
  rhNegativePaedWbA: number;
  rhNegativePaedPcA: number;
  totalA: number;
  percentageOfTotalA: number;
  rhPositiveWbB: number;
  rhPositivePcB: number;
  rhPositivePaedWbB: number;
  rhPositivePaedPcB: number;
  rhNegativeWbB: number;
  rhNegativePcB: number;
  rhNegativePaedWbB: number;
  rhNegativePaedPcB: number;
  totalB: number;
  percentageOfTotalB: number;
  rhPositiveWbAB: number;
  rhPositivePcAB: number;
  rhPositivePaedWbAB: number;
  rhPositivePaedPcAB: number;
  rhNegativeWbAB: number;
  rhNegativePcAB: number;
  rhNegativePaedWbAB: number;
  rhNegativePaedPcAB: number;
  totalAB: number;
  percentageOfTotalAB: number;
  totalRhPositiveWb: number;
  totalRhPositivePc: number;
  totalRhPositivePaedWb: number;
  totalRhPositivePaedPc: number;
  totalRhNegativeWb: number;
  totalRhNegativePc: number;
  totalRhNegativePaedWb: number;
  totalRhNegativePaedPc: number;
  totalTotal: number;
  totalPercentageOfTotal: number;
  rhPositiveWbOcompatibility: number;
  rhPositivePcOcompatibility: number;
  rhPositivePaedWbOcompatibility: number;
  rhPositivePaedPcOcompatibility: number;
  rhNegativeWbOcompatibility: number;
  rhNegativePcOcompatibility: number;
  rhNegativePaedWbOcompatibility: number;
  rhNegativePaedPcOcompatibility: number;
  totalOcompatibility: number;
  percentageOfTotalOcompatibility: number;
  rhPositiveWbAcompatibility: number;
  rhPositivePcAcompatibility: number;
  rhPositivePaedWbAcompatibility: number;
  rhPositivePaedPcAcompatibility: number;
  rhNegativeWbAcompatibility: number;
  rhNegativePcAcompatibility: number;
  rhNegativePaedWbAcompatibility: number;
  rhNegativePaedPcAcompatibility: number;
  totalAcompatibility: number;
  percentageOfTotalAcompatibility: number;
  rhPositiveWbBcompatibility: number;
  rhPositivePcBcompatibility: number;
  rhPositivePaedWbBcompatibility: number;
  rhPositivePaedPcBcompatibility: number;
  rhNegativeWbBcompatibility: number;
  rhNegativePcBcompatibility: number;
  rhNegativePaedWbBcompatibility: number;
  rhNegativePaedPcBcompatibility: number;
  totalBcompatibility: number;
  percentageOfTotalBcompatibility: number;
  rhPositiveWbABcompatibility: number;
  rhPositivePcABcompatibility: number;
  rhPositivePaedWbABcompatibility: number;
  rhPositivePaedPcABcompatibility: number;
  rhNegativeWbABcompatibility: number;
  rhNegativePcABcompatibility: number;
  rhNegativePaedWbABcompatibility: number;
  rhNegativePaedPcABcompatibility: number;
  totalABcompatibility: number;
  percentageOfTotalABcompatibility: number;
  totalRhPositiveWbcompatibility: number;
  totalRhPositivePccompatibility: number;
  totalRhPositivePaedWbcompatibility: number;
  totalRhPositivePaedPccompatibility: number;
  totalRhNegativeWbcompatibility: number;
  totalRhNegativePccompatibility: number;
  totalRhNegativePaedWbcompatibility: number;
  totalRhNegativePaedPccompatibility: number;
  totalTotalcompatibility: number;
  totalPercentageOfTotalcompatibility: number;
  compatsIssues: number;
  compatsOrders: number;
  compatsPercentageSupply_Orders: number;
  ffp1: number;
  plt1: number;
  plt2: number;
  cryo: number;
  paedPacks: number;
  receivedFromAvailable: StockReceivedFrom[];
issuedToAvailable: StockIssuedTo[];
compiledBy: string;
checkedBy: string;

}